Employee backend

+++ server/invoices.lua
@@ -7,14 +7,16 @@ RegisterNetEvent('qb-phone:server:InvoiceHandler')
 -- EVENT HANDLER(S) --
 
 -- Has player paid something this --
---[[AddEventHandler('qb-phone:server:InvoiceHandler', function(paid, amount, player, resource)
-    local src = source
+--[[AddEventHandler('qb-phone:server:InvoiceHandler', function(paid, amount, source, resource)
 
-    if player == src then
-        if paid and resource == GetCurrentResourceName() then
-            if amount >= config.minPayment then
-                -- Do shit
+    if paid and resource == GetCurrentResourceName() then
+        if amount >= config.minPayment then
+            if Config.RenewedBanking then
+                local cid = Player.PlayerData.citizenid
+                local name = ("%s %s"):format(Player.PlayerData.charinfo.firstname, Player.PlayerData.charinfo.lastname)
+                exports['Renewed-Banking']:handleTransaction(cid, "Phone Invoice", amount, "Paid off phone invoice of $"..amount, name, name, "withdraw")
             end
+            -- Do shit
         end
     end
 end)]]
@@ -27,35 +29,27 @@ RegisterNetEvent('qb-phone:server:PayMyInvoice', function(society, amount, invoi
     local Player = QBCore.Functions.GetPlayer(src)
     local SenderPly = QBCore.Functions.GetPlayerByCitizenId(sendercitizenid)
 
-    if SenderPly and Config.BillingCommissions and Config.BillingCommissions[society] then
-        local commission = math.ceil(amount * Config.BillingCommissions[society])
-        SenderPly.Functions.AddMoney('bank', commission)
+    if Player.Functions.RemoveMoney('bank', amount, "paid-invoice") then
+        if SenderPly and Config.BillingCommissions and Config.BillingCommissions[society] then
+            local commission = math.ceil(amount * Config.BillingCommissions[society])
+            SenderPly.Functions.AddMoney('bank', commission)
+        end
+
+        if SenderPly then
+            TriggerClientEvent('qb-phone:client:CustomNotification', SenderPly.PlayerData.source,
+                "Invoice Paid off by " .. SenderPly.PlayerData.charinfo.firstname .. ".",
+                "Recent Invoice of $" .. amount .. " has been paid.",
+                "fas fa-file-invoice-dollar",
+                "#1DA1F2",
+                7500
+            )
+        end
+
+        TriggerClientEvent('qb-phone:client:RemoveInvoiceFromTable', src, invoiceId)
+        TriggerEvent("qb-phone:server:InvoiceHandler", true, amount, src, resource)
+
+        exports.oxmysql:execute('DELETE FROM phone_invoices WHERE id = ?', {invoiceId})
     end
-
-    Player.Functions.RemoveMoney('bank', amount, "paid-invoice")
-
-    if Config.RenewedBanking then
-        local cid = Player.PlayerData.citizenid
-        local name = ("%s %s"):format(Player.PlayerData.charinfo.firstname, Player.PlayerData.charinfo.lastname)
-        exports['Renewed-Banking']:handleTransaction(cid, "Phone Invoice", amount, "Paid off phone invoice of $"..amount, name, name, "withdraw")
-    end
-
-    TriggerEvent("qb-bossmenu:server:addAccountMoney", society, amount)
-
-    if SenderPly then
-        TriggerClientEvent('qb-phone:client:CustomNotification', SenderPly.PlayerData.source,
-            "Invoice Paid off by " .. SenderPly.PlayerData.charinfo.firstname .. ".",
-            "Recent Invoice of $" .. amount .. " has been paid.",
-            "fas fa-file-invoice-dollar",
-            "#1DA1F2",
-            7500
-        )
-    end
-
-    TriggerClientEvent('qb-phone:client:RemoveInvoiceFromTable', src, invoiceId)
-    TriggerEvent("qb-phone:server:InvoiceHandler", true, amount, src, resource)
-
-    exports.oxmysql:execute('DELETE FROM phone_invoices WHERE id = ?', {invoiceId})
 end)
 
 RegisterNetEvent('qb-phone:server:DeclineMyInvoice', function(amount, invoiceId, sendercitizenid, resource)
